+++ CHANGES.rst
@@ -39,6 +39,7 @@ mio 0.0.6.dev
     
 - Implemented ``sum`` builtin.
 - Added ``try`` and ``raise`` builtins. (*``raise`` is not implemented yet...*).
+- Added support for User level Error(s) and implemented ``Exception raise``
 
 
 mio 0.0.5 (2013-10-29)